refacor(Admin): Split adminPage.js and productAPI.js
adminPage.js controle button click
productAPI.js controle API related with product class
