+++ src/main/resources/static/adminPage.js
@@ -1,9 +1,9 @@
+import { addProduct, deleteProduct, editProduct } from './productAPI.js';
+
 const modal = document.getElementById('productModal');
 const closeButton = document.getElementsByClassName('close')[0];
 const mainCheckbox = document.querySelector('table th input[type="checkbox"]');
 const checkboxes = document.querySelectorAll('table td input[type="checkbox"]');
-
-let apiUrl = window.location.origin + '/api/products';
 
 export function addProductbtnOnClick(page) {
     modal.getElementsByTagName('h1')[0].innerText = 'Add a new product';
@@ -13,7 +13,6 @@ export function addProductbtnOnClick(page) {
     );
     modal.style.display = 'flex';
 }
-
 window.addProductbtnOnClick = addProductbtnOnClick;
 
 export function editProductbtnOnClick(id, page) {
@@ -25,7 +24,6 @@ export function editProductbtnOnClick(id, page) {
     );
     modal.style.display = 'flex';
 }
-
 window.editProductbtnOnClick = editProductbtnOnClick;
 
 closeButton.onclick = function () {
@@ -38,29 +36,11 @@ window.onclick = function (event) {
     }
 };
 
-export function addProduct(page) {
-    const name = document.getElementById('productName').value;
-    const price = document.getElementById('productPrice').value;
-    const imageUrl = document.getElementById('productImage').value;
-
-    fetch(`${apiUrl}?name=${name}&price=${price}&imageurl=${imageUrl}`, {
-        method: 'POST',
-        headers: {
-            'Content-Type': 'application/json',
-        },
-    }).then((response) => {
-        if (response.ok) {
-            pagination(page);
-        }
-    });
-}
-
 export function mainCheckboxOnClick() {
     checkboxes.forEach((checkbox) => {
         checkbox.checked = mainCheckbox.checked;
     });
 }
-
 window.mainCheckboxOnClick = mainCheckboxOnClick;
 
 export function checkboxOnClick() {
@@ -73,7 +53,6 @@ export function checkboxOnClick() {
         mainCheckbox.checked = false;
     }
 }
-
 window.checkboxOnClick = checkboxOnClick;
 
 export function deleteCheckedProductsOnClick(page) {
@@ -89,44 +68,4 @@ export function deleteCheckedProductsOnClick(page) {
     });
     mainCheckbox.checked = false;
 }
-
 window.deleteCheckedProductsOnClick = deleteCheckedProductsOnClick;
-
-export function deleteProduct(id, page) {
-    fetch(`${apiUrl}?id=${id}`, {
-        method: 'DELETE',
-    }).then((response) => {
-        if (response.ok) {
-            pagination(page);
-        }
-    });
-}
-
-window.deleteProduct = deleteProduct;
-
-export function editProduct(id, page) {
-    const name = document.getElementById('productName').value;
-    const price = document.getElementById('productPrice').value;
-    const imageUrl = document.getElementById('productImage').value;
-
-    fetch(
-        `${apiUrl}?id=${id}&name=${name}&price=${price}&imageurl=${imageUrl}`,
-        {
-            method: 'PATCH',
-            headers: {
-                'Content-Type': 'application/json',
-            },
-        }
-    ).then((response) => {
-        if (response.ok) {
-            pagination(page);
-        }
-    });
-}
-
-export function pagination(page) {
-    const originalUrl = `${window.location.origin}/admin?page=${page}`;
-    window.location.href = originalUrl;
-}
-
-window.pagination = pagination;
